internal/helpers/dns_helper.go
package helpers

import (
	"fmt"

	"github.com/compose-spec/compose-go/types"
	"github.com/windsor-hotel/cli/internal/config"
	"github.com/windsor-hotel/cli/internal/di"
)

// DNSHelper handles DNS configuration
type DNSHelper struct {
	ConfigHandler config.ConfigHandler
}

// NewDNSHelper creates a new DNSHelper
func NewDNSHelper(di *di.DIContainer) (*DNSHelper, error) {
	cliConfigHandler, err := di.Resolve("cliConfigHandler")
	if err != nil {
		return nil, fmt.Errorf("error resolving cliConfigHandler: %w", err)
	}

	return &DNSHelper{ConfigHandler: cliConfigHandler.(config.ConfigHandler)}, nil
}

// Initialize performs any necessary initialization for the helper.
func (h *DNSHelper) Initialize() error {
	return nil
}

// GetEnvVars returns the environment variables
func (h *DNSHelper) GetEnvVars() (map[string]string, error) {
	return nil, nil
}

// PostEnvExec performs any necessary actions after the environment has been executed
func (h *DNSHelper) PostEnvExec() error {
	return nil
}

// GetComposeConfig returns the compose configuration
func (h *DNSHelper) GetComposeConfig() (*types.Config, error) {
	return nil, nil
}

// WriteConfig writes the configuration to the specified file
func (h *DNSHelper) WriteConfig() error {
	return nil
}

// Ensure DockerHelper implements Helper interface
var _ Helper = (*DNSHelper)(nil)
